[desktop] im => vips
More cleanup is pending, once we give this nightly a go.
