+++ desktop/package.json
@@ -19,7 +19,7 @@
         "postinstall": "electron-builder install-app-deps",
         "lint": "yarn prettier --check --log-level warn . && yarn eslint && yarn tsc",
         "lint-fix": "yarn prettier --write --log-level warn . && yarn eslint && yarn tsc",
-        "prepare": "node scripts/magick.js"
+        "prepare": "node scripts/vips.js"
     },
     "resolutions": {
         "jackspeak": "2.1.1"
